requirements.txt
python-dotenv 
google-genai 
evdev 
opencv-python